+++ myodd/html/TokenStrikeout.cpp
@@ -11,7 +11,7 @@ TokenStrikeout::~TokenStrikeout()
 }
 
 // if this is the token we are looking for.
-bool TokenStrikeout::OnIsToken(const MYODD_CHAR* lpString, int nLen ) const
+bool TokenStrikeout::OnIsToken(const MYODD_CHAR* lpString, __int64 nLen ) const
 {
   if( nLen == 1 )
   {
